+++ fission-cli/library/Fission/CLI/Linking/PIN/Types.hs
@@ -1,0 +1,8 @@
+-- |
+
+module Fission.CLI.Linking.PIN.Types where
+
+import           Fission.Prelude
+
+newtype PIN = PIN Text
+  deriving newtype (Show, Display, Eq, ToJSON, FromJSON)
