+++ fission-core/library/Fission/Web/Auth/DID.hs
@@ -3,19 +3,19 @@ module Fission.Web.Auth.DID (handler) where
 import           Network.Wai
 
 import           Fission.Prelude
- 
-import qualified Fission.Web.Error      as Web.Error
-import qualified Fission.Web.Auth.Error as Auth
 
-import qualified Fission.Web.Auth.Token              as Token
-import qualified Fission.Web.Auth.Token.Bearer.Types as Bearer
+import qualified Fission.Web.Auth.Error                as Auth
+import qualified Fission.Web.Error                     as Web.Error
+
+import qualified Fission.Web.Auth.Token                as Token
+import qualified Fission.Web.Auth.Token.Bearer.Types   as Bearer
 
 import           Fission.Web.Auth.Token.JWT            as JWT
-import qualified Fission.Web.Auth.Token.JWT.Validation as JWT
 import           Fission.Web.Auth.Token.JWT.Resolver   as JWT
+import qualified Fission.Web.Auth.Token.JWT.Validation as JWT
 
-import           Fission.User.DID.Types
 import           Fission.Authorization.ServerDID
+import           Fission.User.DID.Types
 
 -- | Auth handler for registering DIDs
 -- Ensures properly formatted token but *does not check against DB*
@@ -30,8 +30,9 @@ handler ::
   -> m DID
 handler req =
   case Token.get req of
-    Right (Token.Bearer token@(Bearer.Token jwt rawContent)) ->
-      JWT.check rawContent jwt >>= \case
+    Right (Token.Bearer token@(Bearer.Token jwt rawContent)) -> do
+      serverDID <- getServerDID
+      JWT.check serverDID rawContent jwt >>= \case
         Left err -> do
           logWarn $ "Failed registration with token " <> textDisplay token
           Web.Error.throw err
@@ -45,4 +46,3 @@ handler req =
 
     Left err ->
       Web.Error.throw err
-
