+++ fission-core/package.yaml
@@ -1,5 +1,5 @@
 name: fission-core
-version: '2.8.3.0'
+version: '2.9.0.0'
 category: API
 author:
   - Brooklyn Zelenka
